+++ src/core/main.rs
@@ -290,7 +290,8 @@ impl Pushrod {
 
                         eprintln!(
                             "Widget IDs: current={} parent={} children={:?}",
-                            current_widget_id, current_parent_for_widget,
+                            current_widget_id,
+                            current_parent_for_widget,
                             pushrod_window.get_children_of(current_widget_id)
                         );
                     }